line-bot-fastapi/pyproject.toml
[tool.poetry]
name = "line-bot-fastapi"
version = "0.1.0"
description = ""
authors = ["fishwongy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
line-bot-sdk = "^3.10.2"
httpx = "^0.27.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
emoji = "^2.9.0"

pydantic-settings = "^2.2.1"

google-cloud-secret-manager = "^2.16.3"
google-cloud-logging = "^3.8.0"

requests = "^2.31.0"

[build-system]
requires = ["poetry-core>=1.4.1"]
build-backend = "poetry.core.masonry.api"
